+++ slack-gamebot/server.rb
@@ -11,8 +11,11 @@ module SlackGamebot
       client.team = @team
     end
 
-    def restart!(_wait = 1)
-      logger.info "#{team.name}: socket closed, ignoring restart."
+    def restart!(wait = 1)
+      # when an integration is disabled, a live socket is closed, which causes the default behavior of the client to restart
+      # it would keep retrying without checking for account_inactive or such, we want to restart via service which will disable an inactive team
+      logger.info "#{team.name}: socket closed, restarting ..."
+      SlackGamebot::Service.restart! team, self, wait
     end
   end
 end